Small changes to scripts from feedback (#5449)
* remove space at the end of export

* modify env changes

* modify setups

* update prereqs scripts for macos

* modify and cleanup the scripts

* rename files

* cleanup

* update scripts

* changes update

* update

* update Vagrantfile

* update local registry setting

* add comment back

* Add changes to setup scripts

* update kubectl setup script

+++ vagrant-kubernetes-istio/revert_kubectl_config_host.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 
-# Restore kubectl 
-cp ~/.kube/config_old ~/.kube/config
-rm -rf ~/.kube/config_old
+rm -rf ~/.kube/config
+
+# Restore kubectl
+ls ~/.kube/config_old > /dev/null
+if [ $? -eq 0 ]; then
+	cp ~/.kube/config_old ~/.kube/config
+	rm -rf ~/.kube/config_old
+	echo "Your old kube config file has been restored."
+fi
+
+# Restore KUBECONFIG if existed
+export KUBECONFIG=$KUBECONFIG_SAVED
